melanjutkan fitur utama pembimbing

+++ database/migrations/2021_07_24_182758_create_bobot_pemb_laps_table.php
@@ -16,7 +16,7 @@ class CreateBobotPembLapsTable extends Migration
         Schema::create('bobot_pemb_lap', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('id_sub_clo')->unique();
-            $table->integer('bobot');
+            $table->double('bobot');
             $table->timestamps();
 
             $table->foreign('id_sub_clo')->references('id')->on('sub_clo')->onUpdate('cascade')->onDelete('cascade');
